+++ src/deluge/gui/menu_item/osc/type.h
@@ -18,6 +18,7 @@
 #include "definitions_cxx.hpp"
 #include "gui/menu_item/formatted_title.h"
 #include "gui/menu_item/selection.h"
+#include "gui/menu_item/submenu.h"
 #include "gui/ui/sound_editor.h"
 #include "model/song/song.h"
 #include "processing/engines/audio_engine.h"
@@ -25,23 +26,37 @@
 #include "processing/source.h"
 #include "util/comparison.h"
 
+extern deluge::gui::menu_item::Submenu dxMenu;
+
 namespace deluge::gui::menu_item::osc {
 class Type final : public Selection, public FormattedTitle {
 public:
 	Type(l10n::String name, l10n::String title_format_str) : Selection(name), FormattedTitle(title_format_str){};
 	void beginSession(MenuItem* navigatedBackwardFrom) override { Selection::beginSession(navigatedBackwardFrom); }
 
-	void readCurrentValue() override { this->setValue(soundEditor.currentSource->oscType); }
+	bool mayUseDx() { return soundEditor.currentSource->dxPatch != nullptr; }
+
+	void readCurrentValue() override {
+		int32_t rawVal = (int32_t)soundEditor.currentSource->oscType;
+		if (!mayUseDx() && rawVal > (int32_t)OscType::DX7) {
+			rawVal -= 1;
+		}
+		this->setValue(rawVal);
+	}
 	void writeCurrentValue() override {
 
 		OscType oldValue = soundEditor.currentSource->oscType;
 		auto newValue = this->getValue<OscType>();
+		if (!mayUseDx() && (int32_t)newValue >= (int32_t)OscType::DX7) {
+			newValue = (OscType)((int32_t)newValue + 1);
+		}
 
 		auto needs_unassignment = {
 		    OscType::INPUT_L,
 		    OscType::INPUT_R,
 		    OscType::INPUT_STEREO,
 		    OscType::SAMPLE,
+		    OscType::DX7,
 
 		    // Haven't actually really determined if this needs to be here - maybe not?
 		    OscType::WAVETABLE,
@@ -69,27 +84,40 @@ public:
 		    l10n::getView(STRING_FOR_SAW),           //<
 		    l10n::getView(STRING_FOR_ANALOG_SAW),    //<
 		    l10n::getView(STRING_FOR_WAVETABLE),     //<
-		    l10n::getView(STRING_FOR_SAMPLE),        //<
-		    l10n::getView(STRING_FOR_INPUT_LEFT),    //<
-		    l10n::getView(STRING_FOR_INPUT_RIGHT),   //<
-		    l10n::getView(STRING_FOR_INPUT_STEREO),  //<
 		};
-		options[8] = ((AudioEngine::micPluggedIn || AudioEngine::lineInPluggedIn)) //<
-		                 ? l10n::getView(STRING_FOR_INPUT_LEFT)
-		                 : l10n::getView(STRING_FOR_INPUT);
 
 		if (soundEditor.currentSound->getSynthMode() == SynthMode::RINGMOD) {
-			return {options.begin(), options.begin() + kNumOscTypesRingModdable};
+			return options;
 		}
+
+		options.emplace_back(l10n::getView(STRING_FOR_SAMPLE));
+
+		if (mayUseDx()) {
+			options.emplace_back(l10n::getView(STRING_FOR_DX7));
+		}
+
 		if (AudioEngine::micPluggedIn || AudioEngine::lineInPluggedIn) {
-			return {options.begin(), options.begin() + kNumOscTypes};
+			options.emplace_back(l10n::getView(STRING_FOR_INPUT_LEFT));
+			options.emplace_back(l10n::getView(STRING_FOR_INPUT_RIGHT));
+			options.emplace_back(l10n::getView(STRING_FOR_INPUT_STEREO));
 		}
-		return {options.begin(), options.begin() + kNumOscTypes - 2};
+		else {
+			options.emplace_back(l10n::getView(STRING_FOR_INPUT));
+		}
+
+		return options;
 	}
 
 	bool isRelevant(ModControllableAudio* modControllable, int32_t whichThing) override {
 		Sound* sound = static_cast<Sound*>(modControllable);
 		return (sound->getSynthMode() != SynthMode::FM);
+	}
+
+	MenuItem* selectButtonPress() final {
+		if (soundEditor.currentSource->oscType != OscType::DX7) {
+			return NULL;
+		}
+		return (MenuItem*)&dxMenu;
 	}
 };